+++ public.go
@@ -3,7 +3,6 @@ package main
 import (
 	"github.com/dedis/drand/core"
 	"github.com/dedis/drand/net"
-	crypto "github.com/dedis/drand/protobuf/crypto"
 	"github.com/dedis/drand/protobuf/drand"
 	"github.com/nikkolasg/slog"
 	"github.com/urfave/cli"
@@ -86,7 +85,7 @@ func getCokeyCmd(c *cli.Context) error {
 	}
 	ids := getNodes(c)
 	client := core.NewGrpcClientFromCert(defaultManager)
-	var dkey *crypto.Point
+	var dkey *drand.DistKeyResponse
 	var err error
 	for _, id := range ids {
 		dkey, err = client.DistKey(id.Addr, !c.Bool("tls-disable"))
